pyproject.toml
[tool.poetry]
name = "kaggle-tfug-toxic-language"
version = "0.1.0"
description = ""
authors = ["luba <luciano.eq@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = {git = "https://github.com/huggingface/transformers"}
torch = "2.2"
scikit-learn = "1.4"
polars = "0.20.16"
nltk = "3.8"
datasets = "2.18"
accelerate = "^0.28.0"
sentencepiece = "^0.2.0"
packaging = "^24.0"
python-dotenv = "^1.0.1"
openai = "^1.14.3"
lightgbm = {extras = ["scikit-learn"], version = "^4.3.0"}
tenacity = "^8.2.3"
unidecode = "^1.3.8"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
